0.4.1

+++ LGFSwiftTool/LGFExtension/LGF+Int.swift
@@ -13,12 +13,12 @@ public extension Int {
     
     // MARK: - 获取相对于屏幕宽度
     func lgf_W() -> CGFloat {
-        return CGFloat(self * Int(UIDevice.lgf_ScreenW / 375))
+        return CGFloat(CGFloat(self) * CGFloat(UIDevice.lgf_ScreenW / 375.0))
     }
     
     // MARK: - 获取相对于屏幕高度
     func lgf_H() -> CGFloat {
-        return CGFloat(self * Int(UIDevice.lgf_ScreenH / 667))
+        return CGFloat(CGFloat(self) * CGFloat(UIDevice.lgf_ScreenH / 667.0))
     }
     
 }
